mutator now uses jcommander, allowing for custom zygosity and canonicalness heuristics. tests passing.

+++ somatic/src/main/java/org/campagnelab/dl/somatic/intermediaries/SimulationStrategy.java
@@ -12,5 +12,5 @@ public interface SimulationStrategy {
                                                   BaseInformationRecords.SampleInfo germlineSample,
                                                   BaseInformationRecords.SampleInfo otherSample, SimulationCharacteristics sim);
 
-    void setSeed(int seed);
+    void setSeed(long seed);
 }
